% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fd.R
\name{as.fd.fsd.fd}
\alias{as.fd.fsd.fd}
\title{Converting a functional spatial data object into a functional data object}
\usage{
\method{as.fd}{fsd.fd}(x, ...)
}
\arguments{
\item{x}{a functional spatial data object}

\item{...}{currently unused}
}
\description{
Works as one would expect
}
\examples{
\dontrun{
as.fd(x)
}
}
\seealso{
\code{\link{fsd.fd}}
}
\keyword{fsd}
