% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fd.R
\name{as.fsd.fd}
\alias{as.fsd.fd}
\title{Converting a functional data object into a functional spatial data object}
\usage{
as.fsd.fd(x, ...)
}
\arguments{
\item{x}{a functional data object or something that can be converted into
one.}

\item{...}{currently unused}
}
\description{
Works as one would expect
}
\examples{
\dontrun{
as.fsd.fd(x)
}
}
\seealso{
\code{\link{fsd.fd}}
}
\keyword{fsd}
