% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.fsd.fd.R
\name{diff.fsd.fd}
\alias{diff.fsd.fd}
\title{Calculate the differences for a Functional Spatial Data Object}
\usage{
\method{diff}{fsd.fd}(x, lag = 1, differences = 1, ..., dim = 1)
}
\arguments{
\item{x}{the functional spatial data.}

\item{lag}{the lag.}

\item{differences}{the order of the differences.}

\item{...}{currently unused}

\item{dim}{the dimension with respect to which the differences are to be
computed.}
}
\description{
This function is used to calculate the differences of functional data on a
spatial grid.
}
\examples{
data("temp")
plot(diff(temp, dim = 1))
}
\seealso{
\code{\link{fsd.fd}}
}
\keyword{fsd}
