% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.covariance.R
\name{fsd.covariance}
\alias{fsd.covariance}
\title{Estimate autocovariance operators}
\usage{
fsd.covariance(Y = NULL, h, centered = FALSE, na.rm = FALSE,
  unbiased = FALSE)
}
\arguments{
\item{Y}{the functional spatial data. Either an fd object or an array.}

\item{h}{the spatial lag with respect to which the autocovariance is to be
estimated.}

\item{centered}{a boolean indicating whether or not the data is already
centered.}

\item{na.rm}{whether or not missing values should be ignored.}

\item{unbiased}{whether or not to scale the estimator to be unbiased. FALSE
by default. Is not used if na.rm is TRUE.}
}
\value{
the covariance matrix.
}
\description{
This function is used to estimate a covariance operator of stationary
functional spatial data with respect to some lag h.
}
\examples{
\dontrun{
fsd.covariance(Y, h)
}
}
\keyword{fsd}
