% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fd.R
\name{fsd.fd}
\alias{fsd.fd}
\title{Creating a functional spatial data object}
\usage{
fsd.fd(coef = NULL, basisobj = NULL, fdnames = NULL)
}
\arguments{
\item{coef}{an array of arbitrary dimension 1 + r, where the first dimension
corresponds to basis functions.}

\item{basisobj}{a functional basis object defining the basis.}

\item{fdnames}{a list of length r with each member containing the labels for
the dimensions of the data.}
}
\value{
the \code{fsd.fd} object.
}
\description{
This function extends the possibilities of the function fd found in the
package fda to include functional data on a spatial grid of arbitrary
dimension r.
}
\examples{
\dontrun{
fsd.fd(coef, basisobj)
}
}
\keyword{fsd}
