% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.filter.is.unilateral.R
\name{fsd.filter.is.unilateral}
\alias{fsd.filter.is.unilateral}
\title{Check if functional spatial filter is unilateral}
\usage{
fsd.filter.is.unilateral(A, exclude.origin = FALSE)
}
\arguments{
\item{A}{the functional spatial filter.}

\item{exclude.origin}{whether to exclude the origin from the definition of
unilaterality.}
}
\value{
a Boolean indicating whether A is unilateral or not.
}
\description{
This function is used to verify if a functional spatial filter is unilateral.
}
\examples{
\dontrun{
fsd.filter.is.unilateral(A)
}
}
\seealso{
\code{\link{fsd.filter}}
}
\keyword{fsd}
