% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.filter.R
\name{fsd.filter}
\alias{fsd.filter}
\title{Creating a functional spatial filter}
\usage{
fsd.filter(laglist, operatorlist, basisobj = NULL)
}
\arguments{
\item{laglist}{a list of lags.}

\item{operatorlist}{a list of operators, i.e. matrices.}

\item{basisobj}{a functional basis object defining the basis.}
}
\value{
the \code{fsd.filter} object.
}
\description{
This function creates a functional spatial filter
}
\examples{
\dontrun{
fsd.filter(laglist, operatorlist)
}
}
\keyword{fsd}
