% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fourier.inverse.R
\name{fsd.fourier.inverse}
\alias{fsd.fourier.inverse}
\title{Compute the Fourier inverse}
\usage{
fsd.fourier.inverse(evecs, thlist, laglist)
}
\arguments{
\item{evecs}{the eigenvectors to be expanded.}

\item{thlist}{list of frequencies.}

\item{laglist}{list of lags.}
}
\value{
the filter obtained by applying the Fourier inverse on evecs.
}
\description{
This function is used to calculate the Fourier inverse.
}
\examples{
\dontrun{
fsd.fourier.inverse(evecs, thlist, lags)
}
}
\keyword{fsd}
