% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fourier.R
\name{fsd.fourier}
\alias{fsd.fourier}
\title{Compute the Fourier transform}
\usage{
fsd.fourier(covs, laglist, thlist)
}
\arguments{
\item{covs}{the covariances to be transformed}

\item{laglist}{list of lags.}

\item{thlist}{list of frequencies.}
}
\value{
the Fourier transform of the covariances.
}
\description{
This function is used to calculate the Fourier transform
}
\examples{
\dontrun{
fsd.fourier(covs, thlist, lags)
}
}
\keyword{fsd}
