% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.jb.test.R
\name{fsd.jb.test}
\alias{fsd.jb.test}
\title{Perform a Jarque-Bera Test on Normality of Functional Spatial Data}
\usage{
fsd.jb.test(X.spca, Npc = NULL, L = NULL, var.method = "integral")
}
\arguments{
\item{X.spca}{a spectral principal components analysis as performed by
\code{\link{fsd.spca}}, i.e. a list that in particular contains the
estimated SPC scores and the estimated spectral density.}

\item{Npc}{the number of spectral principal components to use for the test.}

\item{L}{the maximum lag to compute the covariance of the scores during the
computation of the autocovariances.}

\item{var.method}{the method that is used to calculate the long-run variance
of the SFPC scores. Either "direct" or "integral".}
}
\value{
A list with components \item{T4}{the test statistic.}
  \item{p.value}{the p-value.} \item{df}{the degrees of freedom.}
  \item{T4.vector}{the vector of the test statistics for each single SPC.}
}
\description{
This function performs a test on normality of some functional spatial data.
}
\details{
To ensure accuracy of numerical integration during the Fourier
  transform, the frequencies of \code{F} should be a suitably dense grid.
}
\examples{
\dontrun{
fsd.jb.test(F, L)
}
}
\seealso{
\code{\link{fsd.spca}}, \code{\link{fsd.spca.cov}}
}
\keyword{fsd}
