% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.norm.R
\name{fsd.norm}
\alias{fsd.norm}
\title{Compute the norm of functional data}
\usage{
fsd.norm(X)
}
\arguments{
\item{X}{the functional spatial data.}
}
\value{
an array of the norms.
}
\description{
This function is used to compute the norm of functional data.
}
\examples{
\dontrun{
fsd.norm(X)
}
}
\keyword{fsd}
