% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.perm.R
\name{fsd.perm}
\alias{fsd.perm}
\title{Permute the dimensions of a Functional Spatial Data Grid}
\usage{
fsd.perm(fsdobj, perm = NULL)
}
\arguments{
\item{fsdobj}{the functional spatial data.}

\item{perm}{the subscript permutation vector, i.e. a permutation of the
integers 1:r.}
}
\description{
This function is used to permute the order of the dimensions of the spatial
grid of some functional spatial data.
}
\examples{
data("temp")
fsd.plot.data(fsd.perm(temp, c(2, 1, 3)))
}
\seealso{
\code{\link{fsd.fd}}
}
\keyword{fsd}
