% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.plot.covariance.R
\name{fsd.plot.covariance}
\alias{fsd.plot.covariance}
\title{Plot the autocovariance operator of functional spatial data}
\usage{
fsd.plot.covariance(
  X,
  basisobj = NULL,
  q = 0,
  gridsize = NULL,
  set.dims = NULL,
  plot.cor = FALSE,
  na.rm = FALSE,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{X}{the functional spatial data. Either an fd object or an array.}

\item{basisobj}{the basis of the functional data.}

\item{q}{the maximal lag to be plotted.}

\item{gridsize}{the resolution of the grid on which the kernel will be
evaluated.}

\item{set.dims}{the coordinates to be fixed. A vector of dimension r. NaN
represents not fixed components.}

\item{plot.cor}{a boolean indicating whether the correlation or the
covariance should be plotted.}

\item{na.rm}{whether or not missing values should be ignored.}

\item{main}{title of the plot.}

\item{xlab}{label for the x-axis.}

\item{ylab}{label for the y-axis.}

\item{...}{other arguments to pass to the plot function}
}
\description{
This function is used to plot the autocovariance operator of functional
spatial data along a two-dimensional plane of lags.
}
\examples{
data("temp")
fsd.plot.covariance(temp, q = 3, plot.cor = TRUE, zlim = c(-1,1))
}
\seealso{
\code{\link{fsd.plot.data}}
}
\keyword{fsd}
