% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.plot.filters.R
\name{fsd.plot.filters}
\alias{fsd.plot.filters}
\title{Plot functional spatial filters}
\usage{
fsd.plot.filters(A, basisobj = NULL, gridsize = NULL, Npc = 3,
  Lmax = 3, set.dims = NULL, main = NULL, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{A}{the filters.}

\item{basisobj}{the basis of the functional data.}

\item{gridsize}{the number of grid points on which to evaluate the functional
data.}

\item{Npc}{number of filters to plot.}

\item{Lmax}{maximum lag to plot.}

\item{set.dims}{the coordinates to be fixed. A vector of dimension r. NaN
represents not fixed components.}

\item{main}{title of the plot.}

\item{xlab}{label for the x-axis.}

\item{ylab}{label for the y-axis.}
}
\description{
This function is used to plot functional spatial filters,  along a
two-dimensional plane of indices.
}
\examples{
\dontrun{
fsd.plot.filters(A)
}
}
\seealso{
\code{\link{plot.fsd.filter}}, \code{\link{fsd.plot.data}},
  \code{\link{fsd.plot.covariance}}
}
\keyword{fsd}
