% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.sfarma.R
\name{fsd.sfarma}
\alias{fsd.sfarma}
\title{Simulate a spatial functional ARMA process}
\usage{
fsd.sfarma(
  n,
  Sigma = NULL,
  ARfilter = NULL,
  MAfilter = NULL,
  burnin = 30,
  basisobj = NULL,
  noise = "normal",
  do.fixed.point = FALSE,
  max.iter = 50,
  eps = 1e-05
)
}
\arguments{
\item{n}{the sample size.}

\item{Sigma}{the covariance matrix of the noise.}

\item{ARfilter}{the functional spatial filter of the autoregressive part.}

\item{MAfilter}{the functional spatial filter of the moving-average part.}

\item{burnin}{the number of grid points to add on each side of the sample to
achieve stationarity.}

\item{basisobj}{the basis of the functional data.}

\item{noise}{the noise distribution to use. Either "normal" or an integer for
the degrees of freedom of a multivariate Student's t-distribution.}

\item{do.fixed.point}{whether to always use a fixed point iteration.}

\item{max.iter}{the maximum number of iterations in the fixed point
iteration.}

\item{eps}{the stopping criterion for the fixed point iteration.}
}
\value{
one sample of the SFARMA process.
}
\description{
This function is used to simulate a sample from a spatial functional ARMA
(SFARMA) process using a normal or t-distributed noise process.
}
\details{
Please note that convergence in the fixed point iteration depends strongly on
the sample size and the structure of the AR filter. For certain choices of
parameters, even more than \code{n} iterations may be needed.
}
\examples{
\dontrun{
fsd.sfarma(n, Sigma, ARfilter, MAfilter, basisobj)
}
}
\seealso{
\code{\link{fsd.fd}}
}
\keyword{fsd}
