% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.spca.filters.R
\name{fsd.spca.filters}
\alias{fsd.spca.filters}
\title{Calculate the Spectral Principal Components Filters Spatial Functional Data}
\usage{
fsd.spca.filters(F, Npc = 1, L = 3)
}
\arguments{
\item{F}{the spectral density.}

\item{Npc}{the number of principal components to be computed.}

\item{L}{the maximum lag for the filters, as an integer or vector of
integers.}
}
\value{
the SPC filters.
}
\description{
This function calculates the SPC filters from a given spectral density
operator.
}
\details{
The eigenvectors used for the calculation of the Fourier expansion
  are oriented such that the sum of their coordinates over the basis of the
  fd object (i.e. the entries in the \code{coefs} array) is a non-negative
  real number. If the basis is not orthonormal, this is done with respect to
  a virtual orthonormal basis in the background.
}
\examples{
\dontrun{
fsd.spca.filters(F)
}
}
\seealso{
\code{\link{fsd.spca}}
}
\keyword{fsd}
