% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.spca.inverse.R
\name{fsd.spca.inverse}
\alias{fsd.spca.inverse}
\title{Reconstruct the original functional data from the filters and the score}
\usage{
fsd.spca.inverse(A, scores, mean.X = NULL)
}
\arguments{
\item{A}{the filters.}

\item{scores}{the scores as an array.}

\item{mean.X}{the mean of X.}
}
\value{
X the reconstructed functional data
}
\description{
This function is used to reconstruct the original functional data from the
filters and the score. It applies the transposed matrices of the filter
\code{A} to the scores.
}
\examples{
\dontrun{
fsd.spca.inverse(A, scores)
}
}
\seealso{
\code{\link{fsd.spca}}, \code{\link{fsd.spca.scores}}
}
\keyword{fsd}
