% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.spca.R
\name{fsd.spca}
\alias{fsd.spca}
\title{Perform Spectral Principal Components Analysis on Spatial Functional Data}
\usage{
fsd.spca(X, freq.res = 100, Npc = 3, L = 3, q = NULL,
  na.ignore = TRUE, return.F = FALSE, only.filters = FALSE)
}
\arguments{
\item{X}{the functional spatial data. Either an fd object or an array.}

\item{freq.res}{the resolution for the computation of the spectral density.}

\item{Npc}{the number of principal components to be computed.}

\item{L}{the maximum lag for the filters. An integer or vector of integers.}

\item{q}{a tuning parameter for the estimation of the the spectral density
operator. An integer or vector of integers.}

\item{na.ignore}{whether to ignore missing data points in the computation of
the scores.}

\item{return.F}{a boolean indicating whether to return the spectral density.}

\item{only.filters}{a boolean indicating whether to compute only the filters,
leaving out the scores.}
}
\value{
A list with components \item{F}{the spectral density.}
  \item{tuning.params}{a list of the used tuning parameters \code{freq.res},
  \code{q} and \code{Lmax}.} \item{filters}{the SPC filters.}
  \item{scores}{the SPC scores.} \item{var}{the theoretical fractions of
  variance explained by each PC.} \item{X.mean}{the mean of X.}
}
\description{
This function performs spectral PCA on functional spatial data on a grid of
dimension r.
}
\details{
This function can be used to compute spectral PCA. By setting
  \code{q = 0}, it can also be used to perform static PCA.

  Setting a higher value for \code{freq.res} increases the runtime
  significantly, but yields a more accurate result because of reduced
  integration errors.
}
\examples{
\dontrun{
fsd.spca(X)
}
}
\seealso{
\code{\link{fsd.spca.inverse}}, \code{\link{fsd.spectral.density}}
}
\keyword{fsd}
