% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.spca.scores.R
\name{fsd.spca.scores}
\alias{fsd.spca.scores}
\title{Compute the Scores for Spatial Functional Data}
\usage{
fsd.spca.scores(X, A, na.ignore = FALSE)
}
\arguments{
\item{X}{the functional spatial data. Either an fd object or an array.}

\item{A}{the filter used to compute the scores.}

\item{na.ignore}{whether missing values in the data X should be ignored.}
}
\value{
the array of the scores.
}
\description{
This function computes the scores. It applies the matrices of the filter
\code{A} to the data \code{X}.
}
\examples{
\dontrun{
fsd.spca.scores(X, A)
}
}
\seealso{
\code{\link{fsd.spca.inverse}}
}
\keyword{fsd}
