% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.spectral.density.R
\name{fsd.spectral.density}
\alias{fsd.spectral.density}
\title{Estimate the spectral density operator}
\usage{
fsd.spectral.density(X, freq, q = NULL, na.rm = FALSE)
}
\arguments{
\item{X}{the functional spatial data. Either an fd object or an array.}

\item{freq}{the spatial frequencies with respect to which the spectral
density is to be estimated. A vector or a list of vectors defining the
grid.}

\item{q}{the size of the window of the kernel used for estimation. An integer
or a vector. For any dimension, the value \code{q = 1} is equivalent to
\code{q = 0} and means that no lags in this direction are taken into
account.}

\item{na.rm}{whether or not missing values should be ignored.}
}
\value{
the spectral density, i.e. a list consisting of \item{operators}{a
  list of the the spectral density operator at different spatial
  frequencies.} \item{freq}{a list of the spatial frequencies.}
  \item{basis}{the basis object of the functional data.}
  \item{estimation.q}{the tuning parameter \code{q} used in the estimation
  process.}
}
\description{
This function is used to estimate the spectral density operator of stationary
functional spatial data on a spatial frequency grid.
}
\details{
The spectral density is estimated employing a  Bartlett kernel on
  the Euclidean norm of the lags. Therefore, the lag \code{q} is the smallest
  lag not to be taken into consideration.
}
\examples{
\dontrun{
fsd.spectral.density(X, freq)
}
}
\keyword{fsd}
