% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.z.plot.R
\name{fsd.z.plot}
\alias{fsd.z.plot}
\title{Plot functional data on a grid}
\usage{
fsd.z.plot(Zlist, positions, basisobj = NULL, gridsize = 50, ...)
}
\arguments{
\item{Zlist}{list of arrays with coefficients.}

\item{positions}{the positions of the single curves in space.}

\item{basisobj}{the basis of the functional data.}

\item{gridsize}{the number of grid points on which to evaluate the curves.}

\item{...}{other arguments to pass to the plot function}
}
\description{
This function plots functional data on a grid.
}
\examples{
\dontrun{
fsd.z.plot(Zlist, positions, basisobj)
}
}
\keyword{fsd}
