% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fd.R
\name{\%/\%.fsd.fd}
\alias{\%/\%.fsd.fd}
\title{Subsetting a functional spatial data object}
\usage{
\method{\%/\%}{fsd.fd}(fsdobj, s)
}
\arguments{
\item{fsdobj}{a functional spatial data object}

\item{s}{an integer or a vector with one entry for every dimension of the
spatial grid.}
}
\description{
Take only every \code{s}-th element
}
\examples{
data("temp")
plot(temp \%/\% 3)
}
\keyword{fsd}
