% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.fsd.fd.R
\name{mean.fsd.fd}
\alias{mean.fsd.fd}
\title{Compute the mean of spatial functional data}
\usage{
\method{mean}{fsd.fd}(x, trim = 0, na.rm = FALSE, ...,
  margins = NULL)
}
\arguments{
\item{x}{the functional spatial data.}

\item{trim}{currently unused}

\item{na.rm}{whether or not missing values should be ignored.}

\item{...}{currently unused}

\item{margins}{the margins that should not be averaged over.}
}
\value{
the mean of x
}
\description{
This function is used to calculate the mean on functional data on a spatial
grid.
}
\examples{
data("temp")

# the mean temperature curve
plot(mean(temp))
# slight variations with respect to the location can be seen
plot(mean(temp, margins = 1:2))
}
\seealso{
\code{\link{center.fsd.fd}}
}
\keyword{fsd}
