% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fsd.fd.R
\name{plot.fsd.fd}
\alias{plot.fsd.fd}
\title{Plot functional spatial data}
\usage{
\method{plot}{fsd.fd}(x, ..., gridsize = NULL, set.dims = NULL)
}
\arguments{
\item{x}{the functional spatial data.}

\item{...}{graphical parameters.}

\item{gridsize}{the number of grid points on which to evaluate the functional
data.}

\item{set.dims}{the coordinates to be fixed. A vector of dimension r. NaN
represents not fixed components.}
}
\description{
This function is used to plot functional spatial data on a grid of dimension
r,  along a two-dimensional plane of indices.
}
\examples{
data("temp")
plot(temp)
}
\seealso{
\code{\link{fsd.plot.data}}
}
\keyword{fsd}
