% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fsd.filter.R
\name{plot.fsd.filter}
\alias{plot.fsd.filter}
\title{Plot functional spatial filters}
\usage{
\method{plot}{fsd.filter}(x, ..., gridsize = NULL, Npc = 3, Lmax = 3,
  set.dims = NULL)
}
\arguments{
\item{x}{the filter.}

\item{...}{graphical parameters.}

\item{gridsize}{the number of grid points on which to evaluate the functional
data.}

\item{Npc}{number of filters to plot.}

\item{Lmax}{maximum lag to plot.}

\item{set.dims}{the coordinates to be fixed. A vector of dimension r. NaN
represents not fixed components.}
}
\description{
This function is used to plot functional spatial filters,  along a
two-dimensional plane of indices.
}
\examples{
\dontrun{
plot.fsd.filter(x)
}
}
\seealso{
\code{\link{fsd.plot.filters}}
}
\keyword{fsd}
