% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fd.R
\name{+.fsd.fd}
\alias{+.fsd.fd}
\title{Arithmetic for functional spatial data objects}
\usage{
\method{+}{fsd.fd}(fsd1, fsd2)
}
\arguments{
\item{fsd1, fsd2}{functional spatial data objects}
}
\description{
Works as one would expect.
}
\examples{
\dontrun{
fsdobj1 + fsdobj2
}
}
\keyword{fsd}
