% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fd.R
\name{/.fsd.fd}
\alias{/.fsd.fd}
\title{Arithmetic for functional spatial data objects}
\usage{
\method{/}{fsd.fd}(fsdobj, scalar)
}
\arguments{
\item{fsdobj}{a functional spatial data object}

\item{scalar}{a scalar number}
}
\description{
Works as one would expect.
}
\examples{
\dontrun{
fsdobj / 2
}
}
\keyword{fsd}
