% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fsd.filter.R
\name{summary.fsd.filter}
\alias{summary.fsd.filter}
\title{Analyse functional spatial filters}
\usage{
\method{summary}{fsd.filter}(object, ..., use.norm = Inf, make.plot = FALSE)
}
\arguments{
\item{object}{the filter.}

\item{...}{currently unused}

\item{use.norm}{which norm to use for the lags.}

\item{make.plot}{a boolean indicating whether to plot the result.}
}
\value{
object list with components \item{balls}{a data frame containing the
  cumulative sums of the squared norms of the filter coefficients up to some
  maximum lag.} \item{zeroplanes}{a matrix containing the sums of the squared
  norms of the filter coefficients lying on the (hyper-) planes with one
  dimension of the lag equals zero, along with the sum over all lags as
  comparison.} \item{abs.sum}{a data frame containing the cumulative sums of
  the norms of the filter coefficients.}
}
\description{
This function is used to analyse functional spatial filters, indicating how
the squared norms lie.
}
\examples{
\dontrun{
summary.fsd.filter(object)
}
}
\seealso{
\code{\link{fsd.plot.filters}}
}
\keyword{fsd}
