% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd.fd.R
\name{*.fsd.fd}
\alias{*.fsd.fd}
\title{Arithmetic for functional spatial data objects}
\usage{
\method{*}{fsd.fd}(scalar, fsdobj)
}
\arguments{
\item{scalar}{a scalar number}

\item{fsdobj}{a functional spatial data object}
}
\description{
Works as one would expect.
}
\examples{
\dontrun{
2 * fsdobj
}
}
\keyword{fsd}
